feat(BlackTurn): 흑돌 턴 구현 및 테스트 재정의
